requirements.txt
PyQt5
flask_sqlalchemy
pymysql
Flask
alembic